Removed @parsed_configs and storing if parsed in object. Fixed leak!
We were storing whether an object had been parsed or not in
@parsed_configs, inside SAXHandler, which was a huge memory leak in case
we wanted to parse really big files.

Created an instance variable in each element to store whether the
element_config was parsed already or not.

It feels a little bit dirty, but it's much better for garbage collection :)

Tests are passing.