Update rails to 6.0
This updates rails test to the current version. Rails now requires
ActiveRecord classes to have a connection when they are created, so we
can use include ActiveModel::Model to test the same thing.  This also
allows the tests to be run on ruby 2.7 without warnings. The previous
rails version was using a bigdecimal API that doesn't exist anymore.